feat(FX-3149): move fair name and logo to page header (#8155)
* feature(FX-3149): fair name/logo moved to header

* refactor(FX-3149): move fair timer to separate component

* refactor(FX-3149): use responsive props instead of Media
