+++ src/v2/Apps/Fair/Components/FairOverview/index.ts
@@ -1,0 +1,1 @@
+export * from "./FairOverview"
