+++ src/v2/Apps/Fair/Components/__tests__/FairHeader.jest.tsx
@@ -1,6 +1,6 @@
-import { FairHeaderFragmentContainer } from "../FairHeader"
 import { graphql } from "react-relay"
 import { setupTestWrapper } from "v2/DevTools/setupTestWrapper"
+import { FairHeaderFragmentContainer } from "../FairHeader"
 
 jest.unmock("react-relay")
 
@@ -16,31 +16,7 @@ const { getWrapper } = setupTestWrapper({
 })
 
 describe("FairHeader", () => {
-  it("displays basic information about the fair", () => {
-    const wrapper = getWrapper({
-      Fair: () => ({
-        name: "Miart 2020",
-        summary: "This is the summary.",
-      }),
-    })
-
-    expect(wrapper.text()).toContain("Miart 2020")
-    expect(wrapper.text()).toContain("This is the summary.")
-  })
-
-  it("displays both the about content and summary", () => {
-    const wrapper = getWrapper({
-      Fair: () => ({
-        about: "This is the about.",
-        summary: "This is the summary.",
-      }),
-    })
-
-    expect(wrapper.text()).toContain("This is the about.")
-    expect(wrapper.text()).toContain("This is the summary.")
-  })
-
-  it("displays all info about the fair", () => {
+  it("displays fair name", () => {
     const wrapper = getWrapper({
       Fair: () => ({
         name: "Miart 2020",
@@ -49,48 +25,6 @@ describe("FairHeader", () => {
     })
 
     expect(wrapper.text()).toContain("Miart 2020")
-  })
-
-  it("renders articles if they are present", () => {
-    const wrapper = getWrapper({
-      Article: () => ({
-        title: "Miart 2020",
-      }),
-    })
-
-    expect(wrapper.text()).toContain("Miart 2020")
-  })
-
-  it("does not render the collection when it is missing", () => {
-    const wrapper = getWrapper({
-      Fair: () => ({ marketingCollections: [] }),
-    })
-
-    expect(wrapper.text()).not.toContain("Curated Highlights")
-    expect(wrapper.text()).not.toContain("Big Artists, Small Sculptures")
-  })
-
-  it("does not render articles when they are missing", () => {
-    const wrapper = getWrapper({
-      ArticleConnection: () => ({ edges: [] }),
-    })
-
-    expect(wrapper.text()).not.toContain("Miart 2020")
-  })
-
-  it("renders the collection when it is present", () => {
-    const wrapper = getWrapper({
-      MarketingCollection: () => ({
-        title: "Big Artists, Small Sculptures",
-      }),
-      FilterArtworksConnection: () => ({
-        counts: { total: 10 },
-      }),
-    })
-
-    expect(wrapper.text()).toContain("Curated Highlights")
-    expect(wrapper.text()).toContain("Big Artists, Small Sculptures")
-    expect(wrapper.text()).toContain("10 works")
   })
 
   it("displays the relevant timing info", () => {
